// src/app/api/shared/agency-profile/agency-profile-load-response.interface.ts
export interface AgencyProfileLoadResponseInterface {
  id: string;
  agency_name: string;
  agency_profile: string;
  agency_reference: string;
  background_color_head_foot: string;
  background_color_main: string;
  background_color_message: string;
  background_color_property_title: string;
  ceo: string;
  administrator: string;
  city: string;
  country: string;
  director: string;
  fax: string;
  foundation_year: string;
  founder: string;
  social_media_links: {
    website: string;
    linkedin: string;
    instagram: string;
    facebook: string;
    twitter: string;
  };
  main_address: string;
  main_email: string;
  main_landline: string;
  main_mobile: string;
  mobiles: string[];
  landlines: string[];
  emails: string[];
  addresses: string[];
  nb_employees: string;
  text_color_head_foot: string;
  text_color_main: string;
  text_color_message: string;
  text_color_property_title: string;
  watermark_on_brochure: number;
  watermark_on_gateway: number;
  watermark_on_website: number;
  whitelisted_domains: string[];
  logo_id: string;
  logo_url: string;
  watermark_id: string;
  watermark_url: string;
  watermark_opacity: number;
  zip_code: string;
  created_by_id: string;
  created_by_name: string;
  creation_datetime: string;
  update_datetime: string;
  updated_by_id: string;
  updated_by_name: string;
}